Rename formatting utils to general utils
