+++ tests/test_utils.py
@@ -1,5 +1,5 @@
 import sae_bench_utils
-import sae_bench_utils.formatting_utils as formatting_utils
+import sae_bench_utils.general_utils as general_utils
 import sae_bench_utils.indexing_utils as indexing_utils
 import sae_bench_utils.testing_utils as testing_utils
 import torch
@@ -20,7 +20,7 @@ def test_average_results():
     expected_output = {"accuracy": 0.825, "loss": 0.45}
 
     # Call the function
-    output = formatting_utils.average_results_dictionaries(results_dict, dataset_names)
+    output = general_utils.average_results_dictionaries(results_dict, dataset_names)
 
     testing_utils.compare_dicts_within_tolerance(output, expected_output, tolerance=1e-6)
 
@@ -56,29 +56,22 @@ def test_get_eval_uuid_uniqueness(num_calls):
     uuids = [sae_bench_utils.get_eval_uuid() for _ in range(num_calls)]
     assert len(set(uuids)) == num_calls, "Not all generated UUIDs are unique"
 
+
 def test_indexing_utils():
     x = torch.arange(40).reshape((2, 20))
     x[0, 10] += 50  # 2nd highest value
     x[0, 11] += 100  # highest value
-    x[1, 1] += (
-        150  # not inside buffer (it's less than 3 from the start of the sequence)
-    )
-    top_indices = indexing_utils.get_k_largest_indices(
-        x, k=2, buffer=3, no_overlap=False
-    )
+    x[1, 1] += 150  # not inside buffer (it's less than 3 from the start of the sequence)
+    top_indices = indexing_utils.get_k_largest_indices(x, k=2, buffer=3, no_overlap=False)
     assert top_indices.tolist() == [[0, 11], [0, 10]]
-    top_indices_no_overlap = indexing_utils.get_k_largest_indices(
-        x, k=2, buffer=3, no_overlap=True
-    )
+    top_indices_no_overlap = indexing_utils.get_k_largest_indices(x, k=2, buffer=3, no_overlap=True)
     assert top_indices_no_overlap.tolist() == [[0, 11], [1, 16]]
 
     # TODO - add test here (should get high values that are still strictly below 10)
     # iw_indices = get_iw_sample_indices(x, k=5, buffer=3, threshold=10)
     # # print(x[iw_indices[:, 0], iw_indices[:, 1]])
 
-    x_top_values_with_context = indexing_utils.index_with_buffer(
-        x, top_indices, buffer=3
-    )
+    x_top_values_with_context = indexing_utils.index_with_buffer(x, top_indices, buffer=3)
     assert x_top_values_with_context[0].tolist() == [
         8,
         9,
